+++ src/main/java/com/dynatrace/openkit/core/objects/NullRootAction.java
@@ -50,6 +50,11 @@ public enum NullRootAction implements RootAction {
     }
 
     @Override
+    public Action reportValue(String valueName, long value) {
+        return this;
+    }
+
+    @Override
     public Action reportValue(String valueName, double value) {
         return this;
     }